Improve accessing types API
